+++ bittensor/__init__.py
@@ -39,15 +39,24 @@ import bittensor.config
 import bittensor.logging
 import bittensor.receptor
 import bittensor.router
-import bittensor.subtensor
 import bittensor.nucleus
 
-from bittensor.axons import axon
-from bittensor.dendrites import dendrite
-from bittensor.clis import cli
-from bittensor.executors import executor
-from bittensor.wallets import wallet
-from bittensor.metagraphs import metagraph
+from bittensor._axon import axon as axon
+from bittensor._dendrite import dendrite as dendrite
+from bittensor._cli import cli as cli
+from bittensor._executor import executor as executor
+from bittensor._wallet import wallet as wallet
+from bittensor._metagraph import metagraph as metagraph
+from bittensor._subtensor import subtensor as subtensor
+
+from bittensor._axon.axon_impl import Axon as Axon
+from bittensor._dendrite.dendrite_impl import Dendrite as Dendrite
+from bittensor._cli.cli_impl import Cli as Cli
+from bittensor._executor.executor_impl import Executor as Executor
+from bittensor._wallet.wallet_impl import Wallet as Wallet
+from bittensor._metagraph.metagraph_impl import Metagraph as Metagraph
+from bittensor._subtensor.subtensor_impl import Subtensor as Subtensor
+
 
 # ---- LOGGING ----
 __debug_on__ = False
