+++ bittensor/_subtensor/__init__.py
@@ -1,0 +1,99 @@
+# The MIT License (MIT)
+# Copyright © 2021 Yuma Rao
+
+# Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated 
+# documentation files (the “Software”), to deal in the Software without restriction, including without limitation 
+# the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, 
+# and to permit persons to whom the Software is furnished to do so, subject to the following conditions:
+
+# The above copyright notice and this permission notice shall be included in all copies or substantial portions of 
+# the Software.
+
+# THE SOFTWARE IS PROVIDED “AS IS”, WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO
+# THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL 
+# THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION 
+# OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER 
+# DEALINGS IN THE SOFTWARE.
+import argparse
+import copy
+import bittensor
+from munch import Munch
+from bittensor.substrate import SubstrateWSInterface
+
+from . import subtensor_impl
+
+custom_type_registry = {
+        "runtime_id": 2,
+        "types": {
+            "NeuronMetadataOf": {
+                "type": "struct",
+                "type_mapping": [["ip", "u128"], ["port", "u16"], ["ip_type", "u8"], ["uid", "u64"], ["modality", "u8"], ["hotkey", "AccountId"], ["coldkey", "AccountId"]]
+            }
+        }
+    }
+
+class subtensor:
+    """
+    Handles interactions with the subtensor chain.
+    """
+
+    def __new__(
+            self, 
+            config: 'Munch' = None,
+            network: str = None,
+            chain_endpoint: str = None
+        ):
+        r""" Initializes a subtensor chain interface.
+            Args:
+                config (:obj:`Munch`, `optional`): 
+                    metagraph.Metagraph.config()
+                network (default='akira', type=str)
+                    The subtensor network flag. The likely choices are:
+                            -- akira (testing network)
+                            -- kusanagi (main network)
+                    If this option is set it overloads subtensor.chain_endpoint with 
+                    an entry point node from that network.
+                chain_endpoint (default=None, type=str)
+                    The subtensor endpoint flag. If set, overrides the --network flag.
+        """
+        if config == None:
+            config = subtensor.default_config()
+        config.subtensor.network = network if network != None else config.subtensor.network
+        config.subtensor.chain_endpoint = chain_endpoint if chain_endpoint != None else config.subtensor.chain_endpoint
+        self.config = copy.deepcopy(config)
+        subtensor.check_config(config)
+        substrate = SubstrateWSInterface(
+            address_type = 42,
+            type_registry_preset='substrate-node-template',
+            type_registry = custom_type_registry,
+        )
+
+        return subtensor_impl.Subtensor( config, substrate )
+
+    @staticmethod
+    def default_config() -> Munch:
+        # Parses and returns a config Munch for this object.
+        parser = argparse.ArgumentParser(); 
+        subtensor.add_args(parser) 
+        config = bittensor.config.Config.to_config(parser); 
+        return config
+    
+    @staticmethod   
+    def add_args(parser: argparse.ArgumentParser):
+        try:
+            parser.add_argument('--subtensor.network', default='kusanagi', type=str, 
+                                help='''The subtensor network flag. The likely choices are:
+                                        -- akira (testing network)
+                                        -- kusanagi (main network)
+                                    If this option is set it overloads subtensor.chain_endpoint with 
+                                    an entry point node from that network.
+                                    ''')
+            parser.add_argument('--subtensor.chain_endpoint', default=None, type=str, 
+                                help='''The subtensor endpoint flag. If set, overrides the --network flag.
+                                    ''')
+        except:
+            pass
+        
+    @staticmethod   
+    def check_config(config: Munch):
+        pass
